components/fincsi_backend/pyproject.toml
[tool.poetry]
name = "fincsi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
djangorestframework = "^3.14"
dj-rest-auth = "^5.0"
django-allauth = "^0.56-dev"
psycopg = {extras = ["binary"], version = "^3.1"}
pillow = "^10.1"
django-rest-swagger = "^2.2.0"
drf-spectacular = "^0.26.5"
# Temporary workaround to get poetry2nix to build it until their version is updated
rpds-py = "0.10.6"
python-dotenv = "^1.0.0"
django-filter = "^23.3"
django-debug-toolbar = "^4.2.0"
django-extensions = "^3.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
